Move actions in spark form to service page actions
